Sparse matrix refactoring.

+++ src/common/Color.h
@@ -31,10 +31,10 @@ struct Color {
 	//inline operator QColor() const { return QColor::fromRgbF(clamp(r, 0.0f, 1.0f), clamp(g, 0.0f, 1.0f), clamp(b, 0.0f, 1.0f), clamp(a, 0.0f, 1.0f)); }
 	inline operator float*() { return &r; }
 	inline uint32_t ToUint32() {
-		uint32_t rr = rint32(clamp<float>(r * 255.0f, 0.0f, 255.0f));
-		uint32_t gg = rint32(clamp<float>(g * 255.0f, 0.0f, 255.0f));
-		uint32_t bb = rint32(clamp<float>(b * 255.0f, 0.0f, 255.0f));
-		uint32_t aa = rint32(clamp<float>(a * 255.0f, 0.0f, 255.0f));
+		uint32_t rr = (uint32_t) rint32(clamp<float>(r * 255.0f, 0.0f, 255.0f));
+		uint32_t gg = (uint32_t) rint32(clamp<float>(g * 255.0f, 0.0f, 255.0f));
+		uint32_t bb = (uint32_t) rint32(clamp<float>(b * 255.0f, 0.0f, 255.0f));
+		uint32_t aa = (uint32_t) rint32(clamp<float>(a * 255.0f, 0.0f, 255.0f));
 		return (aa << 24) | (rr << 16) | (gg << 8) | bb;
 	}
 };
